Changed .gitignore file

+++ noise.py
@@ -50,5 +50,6 @@ def createNoise(scale, precision = 10):
                     matrix[x][y] = 0
                 elif neighborscount == 4 or neighborscount == 3:
                     matrix[x][y] = 1
-    os.remove("Game Files/map.npy")
-    numpy.save("Game Files/map", matrix)
+    if os.path.exists("Game Files=map.npy"):
+        os.remove("Game Files/map.npy")
+    numpy.save("Game Files/map.npy", matrix)
